refactor(xml): replace `XMLInputFactoryFactory` with `HardenedStaxInputFactory` for improved security and consistency
